Move append_file_to_writer to utils
